requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
conan==2.5.0
configobj==5.0.9
contourpy==1.3.1
cryptography==45.0.4
cycler==0.12.1
Cython==0.29.32
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dpath==2.2.0
dulwich==0.22.8
dvc==3.60.1
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.0
fasteners==0.19
fastjsonschema==2.21.1
filelock==3.17.0
flatbuffers==25.2.10
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.56.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.2.0
fst-pso==1.8.1
funcy==2.0
FuzzyTM==2.0.9
gast==0.6.0
gensim==4.3.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
google-play-scraper==1.2.7
grandalf==0.8
grpcio==1.70.0
gto==1.7.2
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.36.0
isoduration==20.11.0
iterative-telemetry==0.0.10
jax==0.5.2
jaxlib==0.5.1
jedi==0.19.2
Jinja2==3.1.5
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kagglehub==0.3.10
keras==3.9.0
kiwisolver==1.4.8
kombu==5.5.4
libclang==18.1.1
llvmlite==0.44.0
lxml==5.4.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
miniful==0.0.6
mistune==3.1.3
ml-dtypes==0.4.1
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.4.4
multitasking==0.0.11
namex==0.0.8
narwhals==1.34.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ngclearn==1.2b3
ngcsimlib==0.3b5
notebook_shim==0.2.4
numba==0.61.2
numpy==1.25.2
omegaconf==2.3.0
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==1.5.3
pandas-datareader==0.10.0
pandas-ta==0.3.14b0
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
patch-ng==1.17.4
patchify==0.2.3
pathspec==0.12.1
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.7
plotly==5.9.0
prometheus_client==0.22.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.0
pyFUME==0.3.1
pygit2==1.18.0
Pygments==2.19.1
pygtrie==2.5.0
PyJWT==2.9.0
pynance==1.0.0
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.1
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.9.3
scmrepo==3.3.11
seaborn==0.13.2
semver==3.0.4
Send2Trash==1.8.3
shap==0.48.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
simpful==2.12.0
six==1.17.0
slicer==0.0.8
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
sqltrie==0.11.2
stack-data==0.6.3
streamlit==1.44.1
sympy==1.13.1
ta-lib @ file:///C:/Users/HP/Downloads/ta_lib-0.6.3-cp310-cp310-win_amd64.whl#sha256=effb277b2470f44e41cd2ac350b2c73b790c6ca1f8f0cd65264f97358f1b8e24
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==1.26.20
vine==5.1.0
voluptuous==0.15.2
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.17.2
xgboost==3.0.2
yarl==1.20.1
yfinance==0.2.55
zc.lockfile==3.0.post1
